Add explicit license terms for LGPL 3 or later
The last version of pyClamd from Alexandre Norman, the original author,
(0.4.0) was released under "LGPL v3 (or later)", but this seems to have been
lost in the various forks over the years.

Re-add this explicit version back, plus a copy of the license, and add an SPDX
compatible license field.